avoid conflicts with other libraries
If another library also require top level libraries with the same name,
then require will not load it again, as it's supposed to be loaded
already. With hash.rb and symbol.rb at top level, being common names for
libraries doing monkey patching, it makes this library unsafe to use.